+++ packages/ts/src/plugin.ts
@@ -1,7 +1,9 @@
 import { createPlugin } from "@flint.fyi/core";
 
+import anyReturns from "./rules/anyReturns.js";
 import asyncPromiseExecutors from "./rules/asyncPromiseExecutors.js";
 import caseDeclarations from "./rules/caseDeclarations.js";
+import caseDuplicates from "./rules/caseDuplicates.js";
 import chainedAssignments from "./rules/chainedAssignments.js";
 import classAssignments from "./rules/classAssignments.js";
 import consecutiveNonNullAssertions from "./rules/consecutiveNonNullAssertions.js";
@@ -10,30 +12,41 @@ import constructorReturns from "./rules/constructorReturns.js";
 import debuggerStatements from "./rules/debuggerStatements.js";
 import defaultCaseLast from "./rules/defaultCaseLast.js";
 import duplicateArguments from "./rules/duplicateArguments.js";
+import elseIfDuplicates from "./rules/elseIfDuplicates.js";
 import emptyDestructures from "./rules/emptyDestructures.js";
 import emptyStaticBlocks from "./rules/emptyStaticBlocks.js";
 import exceptionAssignments from "./rules/exceptionAssignments.js";
+import forDirections from "./rules/forDirections.js";
 import forInArrays from "./rules/forInArrays.js";
 import functionAssignments from "./rules/functionAssignments.js";
+import functionNewCalls from "./rules/functionNewCalls.js";
 import generatorFunctionYields from "./rules/generatorFunctionYields.js";
 import globalAssignments from "./rules/globalAssignments.js";
-import loopAwaits from "./rules/loopAwaits.js";
+import globalObjectCalls from "./rules/globalObjectCalls.js";
 import namespaceDeclarations from "./rules/namespaceDeclarations.js";
 import negativeZeroComparisons from "./rules/negativeZeroComparisons.js";
 import newExpressions from "./rules/newExpressions.js";
 import newNativeNonConstructors from "./rules/newNativeNonConstructors.js";
+import nonOctalDecimalEscapes from "./rules/nonOctalDecimalEscapes.js";
+import numericLiteralParsing from "./rules/numericLiteralParsing.js";
 import objectProto from "./rules/objectProto.js";
 import octalEscapes from "./rules/octalEscapes.js";
 import octalNumbers from "./rules/octalNumbers.js";
 import returnAssignments from "./rules/returnAssignments.js";
+import selfAssignments from "./rules/selfAssignments.js";
+import selfComparisons from "./rules/selfComparisons.js";
+import sequences from "./rules/sequences.js";
 import sparseArrays from "./rules/sparseArrays.js";
 import symbolDescriptions from "./rules/symbolDescriptions.js";
+import typeofComparisons from "./rules/typeofComparisons.js";
 import unassignedVariables from "./rules/unassignedVariables.js";
 import undefinedVariables from "./rules/undefinedVariables.js";
 import unicodeBOMs from "./rules/unicodeBOMs.js";
 import unnecessaryCatches from "./rules/unnecessaryCatches.js";
+import unsafeNegations from "./rules/unsafeNegations.js";
 import variableDeletions from "./rules/variableDeletions.js";
 import voidOperator from "./rules/voidOperator.js";
+import loopAwaits from "./rules/loopAwaits.js";
 
 export const ts = createPlugin({
 	files: {
@@ -41,8 +54,10 @@ export const ts = createPlugin({
 	},
 	name: "ts",
 	rules: [
+		anyReturns,
 		asyncPromiseExecutors,
 		caseDeclarations,
+		caseDuplicates,
 		chainedAssignments,
 		classAssignments,
 		consecutiveNonNullAssertions,
@@ -51,28 +66,39 @@ export const ts = createPlugin({
 		debuggerStatements,
 		defaultCaseLast,
 		duplicateArguments,
+		elseIfDuplicates,
 		emptyDestructures,
 		emptyStaticBlocks,
 		exceptionAssignments,
+		forDirections,
 		forInArrays,
 		functionAssignments,
+		functionNewCalls,
 		generatorFunctionYields,
 		globalAssignments,
 		loopAwaits,
+		globalObjectCalls,
 		namespaceDeclarations,
 		negativeZeroComparisons,
 		newExpressions,
 		newNativeNonConstructors,
+		nonOctalDecimalEscapes,
+		numericLiteralParsing,
 		objectProto,
 		octalEscapes,
 		octalNumbers,
 		returnAssignments,
+		selfAssignments,
+		selfComparisons,
+		sequences,
 		sparseArrays,
 		symbolDescriptions,
+		typeofComparisons,
 		unassignedVariables,
 		undefinedVariables,
 		unicodeBOMs,
 		unnecessaryCatches,
+		unsafeNegations,
 		variableDeletions,
 		voidOperator,
 	],